Remove reference to FirebaseAuth hold by SessionManager
Since this instance can (apparently, see SessionManager.sessionState()) be replaced when the authentication state of the signed-in user changes and the fact that multiple threads can access this instance, it doesn't seem like a good solution to hold a reference to it and update it when the user authentication state changes.

Now, the reference is always taken from FirebaseAuth.getInstance() which is expected to be thread-safe.